changing helpers and networking from js to ts
