+++ src/helpers/Colors.ts
@@ -5,7 +5,7 @@ const colors = {
   warmGrey: "#959595",
   borderGrey: "#f0f0f0",
   whiteFour: "#f8f8f8",
-  strongPink: "#ff197b",
+  strongPink: "#ff197b"
 };
 
 export default colors;
