+++ src/helpers/Strings.ts
